requirements.txt
requests==2.28.0
moviepy==1.0.3
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pymongo==4.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
cloudinary==1.37.0
pika==1.3.1
gtts==2.3.0